# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SmartGreenHouse API specification",
    "description": "dont even read this",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "145.48.6.78:443/service",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/sensornode/register": {
      "post": {
        "tags": ["SensorNode"],
        "summary": "Creating a new sensornode",
        "responses": {
          "200": {
            "description": "Created sensornode",
            "schema": {
              "properties": {
                "sensornode": {
                  "type": "string",
                  "description": "The id of the created sensornode"
                }
              }
            }
          },
          "500": {
            "description": "sensornode creation failed"
          }
        }
      },
      "parameters": [
        {
          "name": "sensornode details",
          "in": "body",
          "required": "true",
          "description": "Creating an new sensornode",
          "schema": {
            "type": "object",
            "required": true,
            "properties": {
              "name": {
                "type": "string"
              },
              "greenhouseDepartment": {
                "type": "string"
              },
              "latitude": {
                "type": "string"
              },
              "longitude": {
                "type": "string"
              },
              "hardwareSerial": {
                "type": "string"
              }
            }
          }
        },
        {
          "name": "x-access-token",
          "in": "header",
          "required": true,
          "type": "string"
        }
      ]
    },
    "/sensornode/update": {
      "post": {
        "tags": ["SensorNode"],
        "summary": "Update sensornode",
        "responses": {
          "200": {
            "description": "updated sensornode",
            "schema": {
              "properties": {
                "greenhousedepartement": {
                  "type": "string",
                  "description": "The id of the updated sensornode"
                }
              }
            }
          },
          "500": {
            "description": "sensornode update failed."
          }
        },
        "description": "Update sensornode",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": true,
            "description": "credentials of sensornode that we want to update",
            "schema": {
              "type": "object",
              "required": true,
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "greenhouseDepartment": {
                  "type": "string"
                },
                "latitude": {
                  "type": "string"
                },
                "longitude": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"]
      }
    },
    "/sensornode/getAll": {
      "get": {
        "tags": ["SensorNode"],
        "summary": "get all sensornodes for greenhousedepartment",
        "responses": {
          "200": {
            "description": "sensornodes",
            "schema": {
              "properties": {
                "greenhouse": {
                  "type": "string",
                  "description": "The ids of the sensornodes"
                }
              }
            }
          },
          "500": {
            "description": "get all failed"
          }
        },
        "description": "get all sensornodes for greenhousedepartment",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "greenhousedepartment",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "produces": ["application/json"]
      }
    },
    "/sensordata/register": {
      "post": {
        "tags": ["SensorData"],
        "summary": "Creating a new sensordata",
        "responses": {
          "200": {
            "description": "Created sensordata",
            "schema": {
              "properties": {
                "sensornode": {
                  "type": "string",
                  "description": "The id of the created sensordata"
                }
              }
            }
          },
          "500": {
            "description": "sensordata creation failed"
          }
        }
      },
      "parameters": [
        {
          "name": "sensordata details",
          "in": "body",
          "required": "true",
          "description": "Creating an new sensordata",
          "schema": {
            "type": "object",
            "required": true,
            "properties": {
              "node": {
                "type": "string"
              },
              "sensorType": {
                "type": "string"
              },
              "value": {
                "type": "string"
              }
            }
          }
        },
        {
          "name": "x-access-token",
          "in": "header",
          "required": true,
          "type": "string"
        }
      ]
    },
    "/sensordata/getAll": {
      "get": {
        "tags": ["SensorData"],
        "summary": "get all sensordata for sensornode id",
        "responses": {
          "200": {
            "description": "sensordata",
            "schema": {
              "properties": {
                "greenhouse": {
                  "type": "string",
                  "description": "The ids of the sensordata"
                }
              }
            }
          },
          "500": {
            "description": "get all failed"
          }
        },
        "description": "get all sensordata for sensornode",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "node",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "produces": ["application/json"]
      }
    },
...
  }
}